Add interfaces for FilmService and UserService. Add new fields for User and Film.
