+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,63 +1,21 @@
 package ru.yandex.practicum.filmorate.service;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Slf4j
-@Service
-public class FilmService {
-    private static int lastId = 1;
-    private final FilmStorage filmStorage;
-    private final UserStorage userStorage;
+public interface FilmService {
+    Film createFilm(Film film);
 
-    @Autowired
-    public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
-        this.filmStorage = filmStorage;
-        this.userStorage = userStorage;
-    }
+    Film updateFilm(Film film);
 
-    public Film createFilm(Film film) {
-        film.setId(generateId());
-        return filmStorage.createFilm(film);
-    }
+    List<Film> getFilms();
 
-    public Film updateFilm(Film film) {
-        return filmStorage.updateFilm(film);
-    }
+    Film getFilmById(int id);
 
-    public List<Film> getFilms() {
-        return filmStorage.getFilms();
-    }
+    void addLike(int filmId, int userId);
 
-    public Film getFilmById(int id) {
-        return filmStorage.getFilmById(id);
-    }
+    void deleteLike(int filmId, int userId);
 
-    public void addLike(int filmId, int userId) {
-        userStorage.getUserById(userId);
-        filmStorage.getFilmById(filmId).getLikes().add(userId);
-    }
-
-    public void deleteLike(int filmId, int userId) {
-        userStorage.getUserById(userId);
-        filmStorage.getFilmById(filmId).getLikes().remove(userId);
-    }
-
-    public List<Film> getTopFilms(int sizeList) {
-        return filmStorage.getFilms().stream()
-                .sorted((f1, f2) -> f2.getLikes().size() - f1.getLikes().size())
-                .limit(sizeList)
-                .collect(Collectors.toList());
-    }
-
-    private int generateId() {
-        return lastId++;
-    }
+    List<Film> getTopFilms(int sizeList);
 }
