+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmServiceImpl.java
@@ -1,0 +1,70 @@
+package ru.yandex.practicum.filmorate.service;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.UserStorage;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Slf4j
+@Service
+public class FilmServiceImpl implements FilmService {
+    private static int lastId = 1;
+    private final FilmStorage filmStorage;
+    private final UserStorage userStorage;
+
+    @Autowired
+    public FilmServiceImpl(FilmStorage filmStorage, UserStorage userStorage) {
+        this.filmStorage = filmStorage;
+        this.userStorage = userStorage;
+    }
+
+    @Override
+    public Film createFilm(Film film) {
+        film.setId(generateId());
+        return filmStorage.createFilm(film);
+    }
+
+    @Override
+    public Film updateFilm(Film film) {
+        return filmStorage.updateFilm(film);
+    }
+
+    @Override
+    public List<Film> getFilms() {
+        return filmStorage.getFilms();
+    }
+
+    @Override
+    public Film getFilmById(int id) {
+        return filmStorage.getFilmById(id);
+    }
+
+    @Override
+    public void addLike(int filmId, int userId) {
+        userStorage.getUserById(userId);
+        filmStorage.getFilmById(filmId).getLikes().add(userId);
+    }
+
+    @Override
+    public void deleteLike(int filmId, int userId) {
+        userStorage.getUserById(userId);
+        filmStorage.getFilmById(filmId).getLikes().remove(userId);
+    }
+
+    @Override
+    public List<Film> getTopFilms(int sizeList) {
+        return filmStorage.getFilms().stream()
+                .sorted((f1, f2) -> f2.getLikes().size() - f1.getLikes().size())
+                .limit(sizeList)
+                .collect(Collectors.toList());
+    }
+
+    private int generateId() {
+        return lastId++;
+    }
+}
